+++ vector/src/main/java/im/vector/app/features/autocomplete/member/AutocompleteMemberController.kt
@@ -10,7 +10,6 @@ package im.vector.app.features.autocomplete.member
 import android.content.Context
 import com.airbnb.epoxy.TypedEpoxyController
 import im.vector.app.features.autocomplete.AutocompleteClickListener
-import im.vector.app.features.autocomplete.autocompleteHeaderItem
 import im.vector.app.features.autocomplete.autocompleteMatrixItem
 import im.vector.app.features.home.AvatarRenderer
 import im.vector.lib.strings.CommonStrings
@@ -43,7 +42,6 @@ class AutocompleteMemberController @Inject constructor(private val context: Cont
         }
         data.forEach { item ->
             when (item) {
-                is AutocompleteMemberItem.Header -> buildHeaderItem(item)
                 is AutocompleteMemberItem.RoomMember -> buildRoomMemberItem(item)
                 is AutocompleteMemberItem.Everyone -> buildEveryoneItem(item)
             }
@@ -53,13 +51,6 @@ class AutocompleteMemberController @Inject constructor(private val context: Cont
     /* ==========================================================================================
      * Helper methods
      * ========================================================================================== */
-
-    private fun buildHeaderItem(header: AutocompleteMemberItem.Header) {
-        autocompleteHeaderItem {
-            id(header.id)
-            title(header.title)
-        }
-    }
 
     private fun buildRoomMemberItem(roomMember: AutocompleteMemberItem.RoomMember) {
         val host = this